cleaning up service and oauth tests
