+++ sample_apps/rso_sample_apps/rust/src/handlers/oauth.rs
@@ -58,73 +58,3 @@ pub async fn handle(
 
     HtmlTemplate(res)
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::config;
-    use httpmock::prelude::*;
-
-    struct ServiceMock {
-        server: MockServer,
-    }
-
-    impl ServiceMock {
-        fn new() -> Self {
-            let server = MockServer::start();
-
-            server.mock(|when, then| {
-                when.method(POST)
-                    .path("/token")
-                    .x_www_form_urlencoded_key_exists("code");
-                then.status(200)
-                    .header("content-type", "application/json")
-                    .body(r#"{"access_token": "xyz", "refresh_token": "abc", "scope": "def", "id_token": "ghi", "token_type": "jkl", "expires_in": 3600}"#);
-            });
-
-            server.mock(|when, then| {
-                when.method(POST).path("/token");
-                then.status(401)
-                    .header("content-type", "application/json")
-                    .body(r#"{}"#);
-            });
-
-            ServiceMock { server }
-        }
-
-        fn configuration(&self) -> config::Configuration {
-            config::Configuration {
-                server: config::Server {
-                    host: "".to_string(),
-                    port: 443,
-                    tls: None,
-                },
-                api_token: "".to_string(),
-                client_id: "".to_string(),
-                client_secret: "".to_string(),
-                provider_url: self.server.url("").to_string(),
-                callback_host: "".to_string(),
-                account_data_url: "".to_string(),
-                champion_data_url: "".to_string(),
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn handle_returns_expected_result() {
-        let cfg = ServiceMock::new().configuration();
-        let filter = handle(&cfg);
-        let res = warp::test::request().path("/oauth?code=abc").reply(&filter);
-
-        assert_eq!(res.await.status(), 200, "Should return 200");
-    }
-
-    #[tokio::test]
-    async fn handle_returns_invalid_code() {
-        let cfg = ServiceMock::new().configuration();
-        let filter = handle(&cfg);
-        let res = warp::test::request().path("/oauth").reply(&filter);
-
-        assert_eq!(res.await.status(), 400, "Should return 400");
-    }
-}
