+++ sample_apps/rso_sample_apps/rust/src/service.rs
@@ -1,7 +1,16 @@
+use crate::config::Configuration;
 use crate::{config, handlers};
 use axum::{routing::get, Router};
 use axum_server::tls_rustls::RustlsConfig;
 use log::{debug, info};
+
+fn create_app(cfg: &Configuration) -> Router {
+    Router::new()
+        .route("/data", get(handlers::data::handle))
+        .route("/oauth", get(handlers::oauth::handle))
+        .route("/", get(handlers::default::handle))
+        .with_state(cfg.clone())
+}
 
 /// Runs the web service
 ///
@@ -19,11 +28,7 @@ use log::{debug, info};
 pub(crate) async fn listen(cfg: &config::Configuration) {
     match cfg.server.addr().parse::<std::net::SocketAddr>() {
         Ok(addr) => {
-            let app = Router::new()
-                .route("/data", get(handlers::data::handle))
-                .route("/oauth", get(handlers::oauth::handle))
-                .route("/", get(handlers::default::handle))
-                .with_state(cfg.clone());
+            let app = create_app(&cfg);
 
             match cfg.clone().server.tls {
                 Some(tls) => {
@@ -48,5 +53,141 @@ pub(crate) async fn listen(cfg: &config::Configuration) {
         Err(e) => {
             panic!("{e}");
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use axum::{
+        body::Body,
+        http::{Request, StatusCode},
+    };
+    use tower::ServiceExt;
+
+    use crate::config;
+    use httpmock::prelude::*;
+
+    struct ServiceMock {
+        server: MockServer,
+    }
+
+    impl ServiceMock {
+        fn new() -> Self {
+            let server = MockServer::start();
+
+            server.mock(|when, then| {
+                when.method(POST)
+                    .path("/token")
+                    .x_www_form_urlencoded_key_exists("code");
+                then.status(200)
+                    .header("content-type", "application/json")
+                    .body(r#"{"access_token": "xyz", "refresh_token": "abc", "scope": "def", "id_token": "ghi", "token_type": "jkl", "expires_in": 3600}"#);
+            });
+
+            server.mock(|when, then| {
+                when.method(POST).path("/token");
+                then.status(401)
+                    .header("content-type", "application/json")
+                    .body(r#"{}"#);
+            });
+
+            ServiceMock { server }
+        }
+
+        fn configuration(&self) -> config::Configuration {
+            config::Configuration {
+                server: config::Server {
+                    host: "".to_string(),
+                    port: 443,
+                    tls: None,
+                },
+                api_token: "".to_string(),
+                client_id: "".to_string(),
+                client_secret: "".to_string(),
+                provider_url: self.server.url("").to_string(),
+                callback_host: "".to_string(),
+                account_data_url: "".to_string(),
+                champion_data_url: "".to_string(),
+            }
+        }
+    }
+
+    #[tokio::test]
+    async fn default() {
+        let cfg = ServiceMock::new().configuration();
+        let app = create_app(&cfg);
+        let response = app
+            .oneshot(Request::builder().uri("/").body(Body::empty()).unwrap())
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), StatusCode::OK);
+    }
+
+    #[tokio::test]
+    async fn oauth_no_code() {
+        let cfg = ServiceMock::new().configuration();
+        let app = create_app(&cfg);
+        let response = app
+            .oneshot(
+                Request::builder()
+                    .uri("/oauth")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), StatusCode::BAD_REQUEST);
+    }
+
+    #[tokio::test]
+    async fn oauth_code() {
+        let cfg = ServiceMock::new().configuration();
+        let app = create_app(&cfg);
+
+        let response = app
+            .oneshot(
+                Request::builder()
+                    .uri("/oauth?code=200")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), StatusCode::OK);
+    }
+
+    #[tokio::test]
+    async fn data_returns_expected_result() {
+        let cfg = ServiceMock::new().configuration();
+        let app = create_app(&cfg);
+
+        let response = app
+            .oneshot(
+                Request::builder()
+                    .uri("/data?access_token=200")
+                    .body(Body::empty())
+                    .unwrap(),
+            )
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), StatusCode::OK);
+    }
+
+    #[tokio::test]
+    async fn data_returns_unauthorized_when_no_access_token() {
+        let cfg = ServiceMock::new().configuration();
+        let app = create_app(&cfg);
+
+        let response = app
+            .oneshot(Request::builder().uri("/data").body(Body::empty()).unwrap())
+            .await
+            .unwrap();
+
+        assert_eq!(response.status(), StatusCode::BAD_REQUEST);
     }
 }
